Update structural-testing.md
SMALL ISSUE REGARDING THE EXERCISE 15

For exercise 15, I think it should be much better to specify that the table must be created in the inverse lexicographic order, to avoid ambiguity.

To fix this, I 've just come with a small note which contains extra information. There is also a possibility to make it an open question and I can write the solution for it. I hope that it is very clear and I will wait for some feedback regarding it.