Update structural-testing.md
SMALL ISSUE REGARDING THE PATH COVERAGE

I've observed that for the example, you have used lazy operators, which can mean that the path can have just 4 cases : 0 x x, 1 0 0, 1 1 x, 1 0 1 (x means don't care value). Moreover, in the video, it is used this expression '' (A && (B | C))'', which gave me that it should be a typo error or maybe a reasoning error. Thus, I think that to avoid any more interpretations, maybe it should be better to not use the lazy operators.

Based on the logical reasoning of that part, I 've just come with a small note which contains some explanations.

SMALL ISSUE REGARDING THE EXERCISE 15

For exercise 15, I think it should be much better to specify that the table must be created in the inverse lexicographic order, to avoid ambiguity.

To fix this, I 've just come with a small note which contains extra information. There is also a possibility to make it an open question and I can write the solution for it.

In conclusion,  I hope that it is very clear and I will wait for some feedback regarding it.