Deleting the wrong example for '.*bug'
Deleting the wrong example for '.*bug'
The example was bad bug. As I understood correctly from the text and within the discussions, if there is the character 'b' followed by a character other than 'u', it goes to the rejecting node and because there is not any transition state out of this state, it will stay there until the input finishes. Thus, I thought the bad bug example was wrong since the first letter is 'b' followed by not 'u'